+++ asterixdb/asterix-external-data/src/test/java/org/apache/asterix/external/classad/Operation.java
@@ -529,7 +529,7 @@ public class Operation extends ExprTree {
         }
 
         // do evaluation
-        sig = privateDoOperation(opKind.getIntegerValue().intValue(), val1, val2, val3, valid1, valid2, valid3, result,
+        sig = privateDoOperation(opKind.getIntegerValue(), val1, val2, val3, valid1, valid2, valid3, result,
                 state, objectPool);
 
         // delete trees which were not significant
@@ -551,13 +551,13 @@ public class Operation extends ExprTree {
 
         // in case of strict operators, if a subexpression is significant and the
         // corresponding value is UNDEFINED or ERROR, propagate only that tree
-        if (isStrictOperator(opKind.getIntegerValue().intValue())) {
+        if (isStrictOperator(opKind.getIntegerValue())) {
             // strict unary operators:  unary -, unary +, !, ~, ()
-            if (opKind.getIntegerValue().intValue() == OpKind_UNARY_MINUS_OP
-                    || opKind.getIntegerValue().intValue() == OpKind_UNARY_PLUS_OP
-                    || opKind.getIntegerValue().intValue() == OpKind_LOGICAL_NOT_OP
-                    || opKind.getIntegerValue().intValue() == OpKind_BITWISE_NOT_OP
-                    || opKind.getIntegerValue().intValue() == OpKind_PARENTHESES_OP) {
+            if (opKind.getIntegerValue() == OpKind_UNARY_MINUS_OP
+                    || opKind.getIntegerValue() == OpKind_UNARY_PLUS_OP
+                    || opKind.getIntegerValue() == OpKind_LOGICAL_NOT_OP
+                    || opKind.getIntegerValue() == OpKind_BITWISE_NOT_OP
+                    || opKind.getIntegerValue() == OpKind_PARENTHESES_OP) {
                 if (val1.isExceptional()) {
                     // the operator is only propagating the value;  only the
                     // subexpression is significant
@@ -565,7 +565,7 @@ public class Operation extends ExprTree {
                 } else {
                     // the node operated on the value; the operator is also
                     // significant
-                    tree.setInnerTree(createOperation(opKind.getIntegerValue().intValue(), t1, objectPool));
+                    tree.setInnerTree(createOperation(opKind.getIntegerValue(), t1, objectPool));
                 }
                 return (true);
             } else {
@@ -582,24 +582,24 @@ public class Operation extends ExprTree {
                     throw new HyracksDataException("Should not reach here");
                 } else {
                     // the node is also significant
-                    tree.setInnerTree(createOperation(opKind.getIntegerValue().intValue(), t1, t2, objectPool));
+                    tree.setInnerTree(createOperation(opKind.getIntegerValue(), t1, t2, objectPool));
                     return (true);
                 }
             }
         } else {
             // non-strict operators
-            if (opKind.getIntegerValue().intValue() == OpKind_IS_OP
-                    || opKind.getIntegerValue().intValue() == OpKind_ISNT_OP) {
+            if (opKind.getIntegerValue() == OpKind_IS_OP
+                    || opKind.getIntegerValue() == OpKind_ISNT_OP) {
                 // the operation is *always* significant for IS and ISNT
-                tree.setInnerTree(createOperation(opKind.getIntegerValue().intValue(), t1, t2, objectPool));
+                tree.setInnerTree(createOperation(opKind.getIntegerValue(), t1, t2, objectPool));
                 return (true);
             }
             // other non-strict binary operators
-            if (opKind.getIntegerValue().intValue() == OpKind_LOGICAL_AND_OP
-                    || opKind.getIntegerValue().intValue() == OpKind_LOGICAL_OR_OP) {
+            if (opKind.getIntegerValue() == OpKind_LOGICAL_AND_OP
+                    || opKind.getIntegerValue() == OpKind_LOGICAL_OR_OP) {
                 if ((SigValues.values()[sig].ordinal() & SigValues.SIG_CHLD1.ordinal()) != 0
                         && (SigValues.values()[sig].ordinal() & SigValues.SIG_CHLD2.ordinal()) != 0) {
-                    tree.setInnerTree(createOperation(opKind.getIntegerValue().intValue(), t1, t2, objectPool));
+                    tree.setInnerTree(createOperation(opKind.getIntegerValue(), t1, t2, objectPool));
                     return (true);
                 } else if ((SigValues.values()[sig].ordinal() & SigValues.SIG_CHLD1.ordinal()) != 0) {
                     tree.setInnerTree(t1);
@@ -613,7 +613,7 @@ public class Operation extends ExprTree {
             }
             // non-strict ternary operator (conditional operator) s ? t : f
             // selector is always significant (???)
-            if (opKind.getIntegerValue().intValue() == OpKind_TERNARY_OP) {
+            if (opKind.getIntegerValue() == OpKind_TERNARY_OP) {
                 Value tmpVal = objectPool.valuePool.get();
                 tmpVal.setUndefinedValue();
                 tree.setInnerTree(Literal.createLiteral(tmpVal, objectPool));
@@ -705,14 +705,14 @@ public class Operation extends ExprTree {
         }
 
         // if splitting is disallowed, fold the value and tree into a tree
-        if (opPtr == null && newOp.getIntegerValue().intValue() != OpKind_NO_OP) {
-            tree.setInnerTree(Operation.createOperation(newOp.getIntegerValue().intValue(), val, tree, objectPool));
+        if (opPtr == null && newOp.getIntegerValue() != OpKind_NO_OP) {
+            tree.setInnerTree(Operation.createOperation(newOp.getIntegerValue(), val, tree, objectPool));
             if (tree.getInnerTree() == null) {
                 return false;
             }
             return true;
         } else if (opPtr != null) {
-            opPtr.setValue(newOp.getIntegerValue().intValue());
+            opPtr.setValue(newOp.getIntegerValue());
         }
         return true;
     }
@@ -726,7 +726,7 @@ public class Operation extends ExprTree {
         if ((tree1.getInnerTree() == null || tree2.getInnerTree() == null)
                 && (tree1.getInnerTree() != null || tree2.getInnerTree() != null)
                 && (op.getIntegerValue() == OpKind_LOGICAL_OR_OP || op.getIntegerValue() == OpKind_LOGICAL_AND_OP)) {
-            privateDoOperation(op.getIntegerValue().intValue(), tree1.getInnerTree() == null ? val1 : dummy,
+            privateDoOperation(op.getIntegerValue(), tree1.getInnerTree() == null ? val1 : dummy,
                     tree2.getInnerTree() == null ? val2 : dummy, dummy, true, true, false, val, objectPool);
             if (val.isBooleanValue()) {
                 tree.setInnerTree(null);
@@ -737,26 +737,26 @@ public class Operation extends ExprTree {
 
         if (tree1.getInnerTree() == null && tree2.getInnerTree() == null) {
             // left and rightsons are only values
-            privateDoOperation(op.getIntegerValue().intValue(), val1, val2, dummy, true, true, false, val, objectPool);
+            privateDoOperation(op.getIntegerValue(), val1, val2, dummy, true, true, false, val, objectPool);
             tree.setInnerTree(null);
             op.setValue(OpKind_NO_OP);
             return true;
         } else if (tree1.getInnerTree() == null
-                && (tree2.getInnerTree() != null && op2.getIntegerValue().intValue() == OpKind_NO_OP)) {
+                && (tree2.getInnerTree() != null && op2.getIntegerValue() == OpKind_NO_OP)) {
             // leftson is a value, rightson is a tree
             tree.setInnerTree(tree2.getInnerTree());
             val.setValue(val1);
             return true;
         } else if (tree2.getInnerTree() == null
-                && (tree1.getInnerTree() != null && op1.getIntegerValue().intValue() == OpKind_NO_OP)) {
+                && (tree1.getInnerTree() != null && op1.getIntegerValue() == OpKind_NO_OP)) {
             // rightson is a value, leftson is a tree
             tree.setInnerTree(tree1.getInnerTree());
             val.setValue(val2);
             return true;
-        } else if ((tree1.getInnerTree() != null && op1.getIntegerValue().intValue() == OpKind_NO_OP)
-                && (tree2.getInnerTree() != null && op2.getIntegerValue().intValue() == OpKind_NO_OP)) {
+        } else if ((tree1.getInnerTree() != null && op1.getIntegerValue() == OpKind_NO_OP)
+                && (tree2.getInnerTree() != null && op2.getIntegerValue() == OpKind_NO_OP)) {
             // left and rightsons are trees only
-            if (null != (newOp = createOperation(op.getIntegerValue().intValue(), tree1, tree2, objectPool))) {
+            if (null != (newOp = createOperation(op.getIntegerValue(), tree1, tree2, objectPool))) {
                 return false;
             }
             tree.setInnerTree(newOp);
@@ -765,25 +765,25 @@ public class Operation extends ExprTree {
         }
 
         // cannot collapse values due to dissimilar ops
-        if ((op1.getIntegerValue().intValue() != OpKind_NO_OP || op2.getIntegerValue().intValue() != OpKind_NO_OP)
+        if ((op1.getIntegerValue() != OpKind_NO_OP || op2.getIntegerValue() != OpKind_NO_OP)
                 && !op.equals(op1) && !op.equals(op1)) {
             // at least one of them returned a value and a tree, and parent does
             // not share the same operation with either child
             ExprTreeHolder newOp1 = objectPool.mutableExprPool.get();
             ExprTreeHolder newOp2 = objectPool.mutableExprPool.get();
 
-            if (op1.getIntegerValue().intValue() != OpKind_NO_OP) {
+            if (op1.getIntegerValue() != OpKind_NO_OP) {
                 newOp1.setInnerTree(
-                        Operation.createOperation(op1.getIntegerValue().intValue(), val1, tree1, objectPool));
+                        Operation.createOperation(op1.getIntegerValue(), val1, tree1, objectPool));
             } else if (tree1.getInnerTree() != null) {
                 newOp1.setInnerTree(tree1.getInnerTree());
             } else {
                 newOp1.setInnerTree(Literal.createLiteral(val1, objectPool));
             }
 
-            if (op2.getIntegerValue().intValue() != OpKind_NO_OP) {
+            if (op2.getIntegerValue() != OpKind_NO_OP) {
                 newOp2.setInnerTree(
-                        Operation.createOperation(op2.getIntegerValue().intValue(), val2, tree2, objectPool));
+                        Operation.createOperation(op2.getIntegerValue(), val2, tree2, objectPool));
             } else if (tree2.getInnerTree() != null) {
                 newOp2.setInnerTree(tree2);
             } else {
@@ -795,7 +795,7 @@ public class Operation extends ExprTree {
                 op.setValue(OpKind_NO_OP);
                 return false;
             }
-            newOp = createOperation(op.getIntegerValue().intValue(), newOp1, newOp2, objectPool);
+            newOp = createOperation(op.getIntegerValue(), newOp1, newOp2, objectPool);
             if (newOp == null) {
                 tree.setInnerTree(null);
                 op.setValue(OpKind_NO_OP);
@@ -809,24 +809,24 @@ public class Operation extends ExprTree {
         if (op.equals(op1) && op.equals(op2)) {
             // same operators on both children . since op!=NO_OP, neither are op1,
             // op2.  so they both make tree and value contributions
-            newOp = createOperation(op.getIntegerValue().intValue(), tree1, tree2, objectPool);
+            newOp = createOperation(op.getIntegerValue(), tree1, tree2, objectPool);
             if (newOp == null) {
                 return false;
             }
-            privateDoOperation(op.getIntegerValue().intValue(), val1, val2, dummy, true, true, false, val, objectPool);
+            privateDoOperation(op.getIntegerValue(), val1, val2, dummy, true, true, false, val, objectPool);
             tree.setInnerTree(newOp);
             return true;
         } else if (op.equals(op1)) {
             // leftson makes a tree,value contribution
             if (tree2.getInnerTree() == null) {
                 // rightson makes a value contribution
-                privateDoOperation(op.getIntegerValue().intValue(), val1, val2, dummy, true, true, false, val,
+                privateDoOperation(op.getIntegerValue(), val1, val2, dummy, true, true, false, val,
                         objectPool);
                 tree.setInnerTree(tree1);
                 return true;
             } else {
                 // rightson makes a tree contribution
-                Operation local_newOp = createOperation(op.getIntegerValue().intValue(), tree1, tree2, objectPool);
+                Operation local_newOp = createOperation(op.getIntegerValue(), tree1, tree2, objectPool);
                 if (local_newOp == null) {
                     tree.setInnerTree(null);
                     op.setValue(OpKind_NO_OP);
@@ -840,13 +840,13 @@ public class Operation extends ExprTree {
             // rightson makes a tree,value contribution
             if (tree1.getInnerTree() == null) {
                 // leftson makes a value contribution
-                privateDoOperation(op.getIntegerValue().intValue(), val1, val2, dummy, true, true, false, val,
+                privateDoOperation(op.getIntegerValue(), val1, val2, dummy, true, true, false, val,
                         objectPool);
                 tree.setInnerTree(tree2);
                 return true;
             } else {
                 // leftson makes a tree contribution
-                Operation local_newOp = createOperation(op.getIntegerValue().intValue(), tree1, tree2, objectPool);
+                Operation local_newOp = createOperation(op.getIntegerValue(), tree1, tree2, objectPool);
                 if (local_newOp == null) {
                     tree.setInnerTree(null);
                     op.setValue(OpKind_NO_OP);
